Revise last change, and made console inputs go to DayplanView not DayplanViewModel
