+++ src/main/java/interface_adapter/DayplanViewModel.java
@@ -7,24 +7,15 @@ public class DayplanViewModel extends ViewModel {
     public final String prompt = "Describe how your day is in a sentence:";
     public String userInput;
 
-
     public DayplanViewModel() {
         super("Dayplan VM");
     }
 
-    /**
-     * Gets String userPrompt from DayplanViewModel
-     * */
-    public String getUserInput() {
-        return this.userInput;
+    public String getTitleLabel() {
+        return this.titleLabel;
     }
 
-    /**
-     * Set the userInput from console
-     * */
-    public void setUserInput() {
-        System.out.println(prompt);
-        Scanner input = new Scanner(System.in);
-        this.userInput = input.nextLine();
+    public String getPrompt() {
+        return this.prompt;
     }
 }
